+++ ipinfo/handler_async.py
@@ -1,5 +1,5 @@
 """
-Main API client handler for fetching data from the IPinfo service.
+Main API client asynchronous handler for fetching data from the IPinfo service.
 """
 
 from ipaddress import IPv4Address, IPv6Address
@@ -7,16 +7,16 @@ import json
 import os
 import sys
 
-import requests
+import aiohttp
 
 from .cache.default import DefaultCache
 from .details import Details
 from .exceptions import RequestQuotaExceededError
 
 
-class Handler:
+class AsyncHandler:
     """
-    Allows client to request data for specified IP address.
+    Allows client to request data for specified IP address asynchronously.
     Instantiates and maintains access to cache.
     """
 
@@ -41,6 +41,9 @@ class Handler:
         if "timeout" not in self.request_options:
             self.request_options["timeout"] = self.REQUEST_TIMEOUT_DEFAULT
 
+        # setup aiohttp
+        self.httpsess = None
+
         # setup cache
         if "cache" in kwargs:
             self.cache = kwargs["cache"]
@@ -52,14 +55,66 @@ class Handler:
                 cache_options["ttl"] = self.CACHE_TTL
             self.cache = DefaultCache(**cache_options)
 
-    def getDetails(self, ip_address=None):
+    async def init(self):
+        """
+        Initializes internal aiohttp connection pool.
+
+        This isn't _required_, as the pool is initialized lazily when needed.
+        But in case you require non-lazy initialization, you may await this.
+
+        This is idempotent.
+        """
+        await self._ensure_aiohttp_ready()
+
+    async def deinit(self):
+        """
+        Deinitialize the async handler.
+
+        This is required in case you need to let go of the memory/state
+        associated with the async handler in a long-running process.
+
+        This is idempotent.
+        """
+        if self.httpsess:
+            await self.httpsess.close()
+            self.httpsess = None
+
+    async def getDetails(self, ip_address=None):
         """Get details for specified IP address as a Details object."""
-        raw_details = self._requestDetails(ip_address)
+        self._ensure_aiohttp_ready()
+
+        # If the supplied IP address uses the objects defined in the built-in
+        # module ipaddress, extract the appropriate string notation before
+        # formatting the URL.
+        if isinstance(ip_address, IPv4Address) or isinstance(
+            ip_address, IPv6Address
+        ):
+            ip_address = ip_address.exploded
+
+        if ip_address in self.cache:
+            return Details(self.cache[ip_address])
+
+        # not in cache; do http req
+        url = self.API_URL
+        if ip_address:
+            url += "/" + ip_address
+        headers = self._get_headers()
+        async with self.httpsess.get(url, headers=headers) as resp:
+            if resp.status == 429:
+                raise RequestQuotaExceededError()
+            resp.raise_for_status()
+            raw_details = await resp.json()
+
+        # format & cache
         self._format_details(raw_details)
+        self.cache[ip_address] = raw_details
+
         return Details(raw_details)
 
-    def getBatchDetails(self, ip_addresses):
+    async def getBatchDetails(self, ip_addresses):
         """Get details for a batch of IP addresses at once."""
+        self._ensure_aiohttp_ready()
+
         result = {}
 
         # Pre-populate with anything we've got in the cache, and keep around
@@ -79,57 +134,40 @@ class Handler:
             else:
                 lookup_addresses.append(ip_address)
 
-        # Do the lookup
+        # all in cache - return early.
+        if len(lookup_addresses) == 0:
+            return result
+
+        # do http req
         url = self.API_URL + "/batch"
         headers = self._get_headers()
         headers["content-type"] = "application/json"
-        response = requests.post(
-            url, json=lookup_addresses, headers=headers, **self.request_options
-        )
-        if response.status_code == 429:
-            raise RequestQuotaExceededError()
-        response.raise_for_status()
+        async with self.httpsess.post(
+            url, data=json.dumps(lookup_addresses), headers=headers
+        ) as resp:
+            if resp.status == 429:
+                raise RequestQuotaExceededError()
+            resp.raise_for_status()
+            json_resp = await resp.json()
 
-        # Fill up cache
-        json_response = response.json()
-        for ip_address, details in json_response.items():
-            self.cache[ip_address] = details
+        # format & fill up cache
+        for ip_address, details in json_resp.items():
+            if isinstance(details, dict):
+                self._format_details(details)
+                self.cache[ip_address] = details
 
-        # Merge cached results with new lookup
-        result.update(json_response)
-
-        # Format every result
-        for detail in result.values():
-            if isinstance(detail, dict):
-                self._format_details(detail)
+        # merge cached results with new lookup
+        result.update(json_resp)
 
         return result
 
-    def _requestDetails(self, ip_address=None):
-        """Get IP address data by sending request to IPinfo API."""
+    def _ensure_aiohttp_ready(self):
+        """Ensures aiohttp internal state is initialized."""
+        if self.httpsess:
+            return
 
-        # If the supplied IP address uses the objects defined in the built-in
-        # module ipaddress extract the appropriate string notation before
-        # formatting the URL.
-        if isinstance(ip_address, IPv4Address) or isinstance(
-            ip_address, IPv6Address
-        ):
-            ip_address = ip_address.exploded
-
-        if ip_address not in self.cache:
-            url = self.API_URL
-            if ip_address:
-                url += "/" + ip_address
-
-            response = requests.get(
-                url, headers=self._get_headers(), **self.request_options
-            )
-            if response.status_code == 429:
-                raise RequestQuotaExceededError()
-            response.raise_for_status()
-            self.cache[ip_address] = response.json()
-
-        return self.cache[ip_address]
+        timeout = aiohttp.ClientTimeout(total=self.request_options["timeout"])
+        self.httpsess = aiohttp.ClientSession(timeout=timeout)
 
     def _get_headers(self):
         """Built headers for request to IPinfo API."""