+++ scripts/fmt.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+DIR=`dirname $0`
+
+# Format the project.
+
+black -l 79 $DIR/../ipinfo $DIR/../tests