+++ tests/handler_async_test.py
@@ -1,25 +1,27 @@
-from ipaddress import IPv4Address
-import json
 import os
 
 from ipinfo.cache.default import DefaultCache
 from ipinfo.details import Details
-from ipinfo.handler import Handler
+from ipinfo.handler_async import AsyncHandler
 import pytest
 
 
-def test_init():
+@pytest.mark.asyncio
+async def test_init():
     token = "mytesttoken"
-    handler = Handler(token)
+    handler = AsyncHandler(token)
     assert handler.access_token == token
     assert isinstance(handler.cache, DefaultCache)
-    assert "US" in handler.countries
+    assert "PK" in handler.countries
+    await handler.deinit()
 
 
-def test_headers():
+@pytest.mark.asyncio
+async def test_headers():
     token = "mytesttoken"
-    handler = Handler(token)
+    handler = AsyncHandler(token)
     headers = handler._get_headers()
+    await handler.deinit()
 
     assert "user-agent" in headers
     assert "accept" in headers
@@ -27,10 +29,11 @@ def test_headers():
 
 
 @pytest.mark.parametrize("n", range(5))
-def test_get_details(n):
+@pytest.mark.asyncio
+async def test_get_details(n):
     token = os.environ.get("IPINFO_TOKEN", "")
-    handler = Handler(token)
-    details = handler.getDetails("8.8.8.8")
+    handler = AsyncHandler(token)
+    details = await handler.getDetails("8.8.8.8")
     assert isinstance(details, Details)
     assert details.ip == "8.8.8.8"
     assert details.hostname == "dns.google"
@@ -78,13 +81,16 @@ def test_get_details(n):
         assert domains["total"] == 12988
         assert len(domains["domains"]) == 5
 
+    await handler.deinit()
+
 
 @pytest.mark.parametrize("n", range(5))
-def test_get_batch_details(n):
+@pytest.mark.asyncio
+async def test_get_batch_details(n):
     token = os.environ.get("IPINFO_TOKEN", "")
-    handler = Handler(token)
+    handler = AsyncHandler(token)
     ips = ["1.1.1.1", "8.8.8.8", "9.9.9.9"]
-    details = handler.getBatchDetails(ips)
+    details = await handler.getBatchDetails(ips)
 
     for ip in ips:
         assert ip in details
@@ -99,53 +105,4 @@ def test_get_batch_details(n):
             assert "abuse" in d
             assert "domains" in d
 
-
-def test_builtin_ip_types():
-    handler = Handler()
-    fake_details = {"country": "US", "ip": "127.0.0.1", "loc": "12.34,56.78"}
-
-    handler._requestDetails = lambda x: fake_details
-
-    details = handler.getDetails(IPv4Address(fake_details["ip"]))
-    assert isinstance(details, Details)
-    assert details.country == fake_details["country"]
-    assert details.country_name == "United States"
-    assert details.ip == fake_details["ip"]
-    assert details.loc == fake_details["loc"]
-    assert details.longitude == "56.78"
-    assert details.latitude == "12.34"
-
-
-def test_json_serialization():
-    handler = Handler()
-    fake_details = {
-        "asn": {
-            "asn": "AS20001",
-            "domain": "twcable.com",
-            "name": "Time Warner Cable Internet LLC",
-            "route": "104.172.0.0/14",
-            "type": "isp",
-        },
-        "city": "Los Angeles",
-        "company": {
-            "domain": "twcable.com",
-            "name": "Time Warner Cable Internet LLC",
-            "type": "isp",
-        },
-        "country": "US",
-        "country_name": "United States",
-        "hostname": "cpe-104-175-221-247.socal.res.rr.com",
-        "ip": "104.175.221.247",
-        "loc": "34.0293,-118.3570",
-        "latitude": "34.0293",
-        "longitude": "-118.3570",
-        "phone": "323",
-        "postal": "90016",
-        "region": "California",
-    }
-
-    handler._requestDetails = lambda x: fake_details
-
-    details = handler.getDetails(fake_details["ip"])
-    assert isinstance(details, Details)
-    assert json.dumps(details.all)
+    await handler.deinit()